Make ObjectsAreEqual{Values} private
This is a breaking change, but these methods are not assertions
(they don't call Fail) although their names look like such,
so I think they should not be exposed.

Alternatively:
- Actually make them assertions?
- Some way to mark them as depracated?

For example today you could write the following in your tests:

assert.ObjectsAreEqual(t, 0.1, 0)

and it would always pass which is kind of confusing.

Also, these methods are not exposed in the Assertions* API.